feat: added /v1/alarms api endpoint

+++ domain/service/alarmService.go
@@ -9,10 +9,11 @@ import (
 
 // AlarmService ...
 type AlarmService interface {
+	Create(memberID, roomID uint, code vo.TaskCode, roomName, diaryTitle, authorNickname string) (*entity.Alarm, error)
 	GetAll(memberID uint) (*entity.Alarms, error)
-	PushByID(memberID uint, al *vo.Alarm) error
-	PushByEmail(email string, al *vo.Alarm) error
-	BroadCast(memberIDs []uint, al *vo.Alarm) error
+	PushByID(memberID uint, al *entity.Alarm) error
+	PushByEmail(email string, al *entity.Alarm) error
+	BroadCast(memberIDs []uint, al *entity.Alarm) error
 }
 
 type alarmService struct {
@@ -29,7 +30,20 @@ func NewAlarmService(ms MemberService, mdr repository.MemberDeviceRepository, ar
 		alarmRepository:        ar,
 	}
 }
-func (as *alarmService) PushByID(memberID uint, al *vo.Alarm) (err error) {
+
+func (as *alarmService) Create(memberID, roomID uint, code vo.TaskCode, roomName, diaryTitle, authorNickname string) (*entity.Alarm, error) {
+	alarm, err := as.alarmRepository.Create(entity.NewAlarm(memberID, roomID, code, roomName, diaryTitle, authorNickname))
+	if err != nil {
+		return nil, err
+	}
+	return alarm, nil
+}
+
+func (as *alarmService) GetAll(memberID uint) (*entity.Alarms, error) {
+	return as.alarmRepository.GetAll(memberID)
+}
+
+func (as *alarmService) PushByID(memberID uint, al *entity.Alarm) (err error) {
 	var deviceTokens []string
 	if deviceTokens, err = as.memberDeviceRepository.GetAllTokens(memberID); err != nil {
 		return
@@ -39,14 +53,16 @@ func (as *alarmService) PushByID(memberID uint, al *vo.Alarm) (err error) {
 	if failedTokens, err = firebase.GetClient().Push(deviceTokens, al); err != nil {
 		return
 	}
-
-	if err = as.memberDeviceRepository.DeleteBatch(failedTokens); err != nil {
-		return
+	if len(failedTokens) > 0 {
+		if err = as.memberDeviceRepository.DeleteBatch(failedTokens); err != nil {
+			return
+		}
 	}
+
 	return
 }
 
-func (as *alarmService) PushByEmail(email string, al *vo.Alarm) (err error) {
+func (as *alarmService) PushByEmail(email string, al *entity.Alarm) (err error) {
 	var member *entity.Member
 	if member, err = as.memberService.GetByEmail(email); err != nil {
 		return
@@ -54,7 +70,7 @@ func (as *alarmService) PushByEmail(email string, al *vo.Alarm) (err error) {
 	return as.PushByID(member.ID, al)
 }
 
-func (as *alarmService) BroadCast(memberIDs []uint, al *vo.Alarm) (err error) {
+func (as *alarmService) BroadCast(memberIDs []uint, al *entity.Alarm) (err error) {
 	var deviceTokens []string
 	if deviceTokens, err = as.memberDeviceRepository.GetAllMemberTokens(memberIDs); err != nil {
 		return
@@ -64,10 +80,10 @@ func (as *alarmService) BroadCast(memberIDs []uint, al *vo.Alarm) (err error) {
 	if failedTokens, err = firebase.GetClient().Push(deviceTokens, al); err != nil {
 		return
 	}
-
-	if err = as.memberDeviceRepository.DeleteBatch(failedTokens); err != nil {
-		return
+	if len(failedTokens) > 0 {
+		if err = as.memberDeviceRepository.DeleteBatch(failedTokens); err != nil {
+			return
+		}
 	}
-
 	return
 }
